Fallback info and computeRoutes response docs
